+++ src/ccbr_tools/hooks/__init__.py
@@ -2,4 +2,5 @@
 Pre-commit hooks for CCBR Bioinformatics Software, intended for use with the pre-commit framework (<https://pre-commit.com/>).
 
 - [detect-absolute-paths](`~ccbr_tools.hooks.detect_absolute_paths`)
+- [sync-nextflow-version](`~ccbr_tools.hooks.sync_nextflow_version`)
 """